redmine63916 修改测试组bug
git-svn-id: http://192.168.1.20:8000/svn/17portal/ec-portal/LTE_PUBLISH@66684 03d26769-fb22-2544-a5c1-c2b74650ae0b

+++ lte-portal-4G/src/main/java/com/al/lte/portal/controller/crm/OrderController.java
@@ -1318,7 +1318,7 @@ public class OrderController extends BaseController {
                                     if(SysConstant.ON.equals(sessionStaff.getPoingtType())){
                                     	 String boActionType = (String) item.get("boActionType");
                                     	 if(item.get("objId") != null){
-                                    		 int objId = ((Integer)item.get("objId")).intValue();//业务对象ID
+                                    		 int objId = (Integer)item.get("objId");//业务对象ID
                                              if("14".equals(boActionType) || objId==13409281){//补换卡   //国际及港澳台漫游电话（包含语音及短信）
                                              	checkFlag = true;
                                              }
